Add files via upload
Conceptual creation of and AI model using Keras on the MNIST dataset. The data is split into a train set and test set, normalized, then used to train the model. From here the model is saved, converted to  TensorFlow Lite model, then tested. Included is a report on my work, including issues encountered and what I did to troubleshoot and fix them